+++ packages/blueprint_lib/src/types.ts
@@ -1,36 +1,8 @@
-import { VineValidator } from '@vinejs/vine'
-import { InferInput } from '@vinejs/vine/types'
+import type { VineValidator } from '@vinejs/vine'
+import type { InferInput } from '@vinejs/vine/types'
 import type { MultipartFile } from '@adonisjs/bodyparser/types'
 import { Constructor } from '@adonisjs/core/types/container'
 
-export type EndpointIO = {
-  output: any
-  input: any
-}
-
-export type Endpoint<IO extends EndpointIO> = {
-  url: (options?: {
-    params?: 'input' extends keyof IO
-      ? 'params' extends keyof IO['input']
-        ? IO['input']['params']
-        : undefined
-      : undefined
-    query?: 'input' extends keyof IO
-      ? 'query' extends keyof IO['input']
-        ? IO['input']['query']
-        : undefined
-      : undefined
-  }) => string
-  method: string
-  form?: boolean
-  io: IO
-}
-
-export type ApiEndpoints = Record<string, Endpoint<EndpointIO>>
-
-export type EndpointKeys<Endpoints extends ApiEndpoints> = keyof Endpoints
-
-// ===============
 // https://github.com/Julien-R44/tuyau/blob/main/packages/utils/src/types.ts
 
 type Prettify<T> = {