+++ packages/castle/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @folie/castle
 
+## 0.0.15
+
+### Patch Changes
+
+- Updated dependencies
+  - @folie/lib@0.0.6
+
 ## 0.0.14
 
 ### Patch Changes