+++ packages/cobalt/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @folie/cobalt
 
+## 0.0.28
+
+### Patch Changes
+
+- Updated dependencies
+  - @folie/blueprint-lib@0.0.7
+
 ## 0.0.27
 
 ### Patch Changes