+++ packages/gate_tan/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @folie/gate-tan
 
+## 0.0.2
+
+### Patch Changes
+
+- dependency fix
+- Updated dependencies
+  - @folie/gate@0.0.10
+  - @folie/lib@0.0.6
+
 ## 0.0.1
 
 ### Patch Changes