+++ packages/lib/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @folie/lib
 
+## 0.0.6
+
+### Patch Changes
+
+- dependency fix
+
 ## 0.0.5
 
 ### Patch Changes