+++ playground/frontend/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @folie/playground-frontend
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @folie/blueprint-lib@0.0.7
+  - @folie/gate-next@0.0.2
+  - @folie/gate-tan@0.0.2
+  - @folie/gate@0.0.10
+  - @folie/lib@0.0.6
+  - @folie/cobalt@0.0.28
+  - @folie/playground-backend@0.0.4
+
 ## 0.0.3
 
 ### Patch Changes